Consistent naming of plugin main class, argument class, and injector config for PomAnalyzer and MavenCrawler
